requirements.txt
pandas
sqlalchemy
psycopg2-binary
python-dotenv
prophet
xgboost
requests
beautifulsoup4
matplotlib
seaborn
